Release 0.6.0
New X-Rate-Limit response header is now included with a value of USED/TOTAL so if you have used 45 of your 200 requests in an hour, the header will look like this:
HTTP/1.1 200 OK
X-Rate-Limit: 45/200
...
New support fields on transaction write endpoints:
flag_color - The transaction color flag
payee_name - If specified and payee_id is not specified, payee_name will be used to resolve an existing payee or create a new payee with this name when creating or updating a transaction.
New account fields
cleared_balance
uncleared_balance
Scheduled transaction dates
date field renamed to date_first which is the first date it was scheduled to fire
New date_next field which is the next date it is scheduled to fire
Account and category balances will now be updated after creating / updating transactions
Payee auto-categorization now supported when specifying a payee_id but not a category_id for transaction create / update. Just like on the web or mobile apps, if payee / category mapping is specified, we will use it.