fix some incorrect parts of the swagger spec
in particular, the spec marks a lot of fields as required when the api
can definitely return `null` for those fields (as determined by making
test requests to the api). i'm not particularly confident that i found
all of the cases of these, but this is just what i ran into when trying
to implement some api clients.

additionally, the spec file claimed that creating transactions and
updating transactions could both use the `SaveTransaction` model, but
this isn't accurate, since updating transactions requires that you pass
in the `id` field. i addressed this by adding a new `UpdateTransaction`
model which does contain an `id` field.